beginning of a major refactoring
Starting with this release, the svelteplot marks are no longer
responsible for scaling the data variables. Instead, the central
<Mark> component is doing this.

This is a necessary refactoring for enabling transforms that are
applied after the data has been scaled, such as the dodge transform.
In theory, the dx and dy "channels" can be considered a post-scale
transformation, already, but like the scaling, they were applied within
each geometry mark class. This resulted in a lot of duplicated code,
and was prone to bugs.

Likewise, the filter transform and the facet assignment had to be
handled by each geom mark prior to this change. Not the facet-filtering
and the filter transform are applied in one place only: the Mark
component.